chore(database): move module into its own folder
